重构了登录的业务逻辑,分离相对应代码

+++ src/main/java/com/suny/association/mapper/LoginTicketMapper.java
@@ -7,6 +7,7 @@ import com.suny.association.pojo.po.LoginTicket;
  * Created by 孙建荣 on 17-9-20.上午8:55
  */
 public interface LoginTicketMapper extends IMapper<LoginTicket> {
-    LoginTicket selectByAccountId(int accountId);
+
+    LoginTicket selectByAccountId(long accountId);
     LoginTicket selectByTicket(String ticket);
 }
